Support named tracks and standalone reminders.
For either of these to work, the top-level item in your json file will
be a dictionary.

A named track is a key with a list as a value. Named tracks work like
the old system: you get one reminder each day.

A standalone reminder is an independent reminder you get once every n
days. For these reminders, the key is the reminder string and the value
is n, how often you want to be reminded.